requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Graph Neural Networks
# Note: torch-geometric extensions will be installed separately (see install.sh)
torch-geometric>=2.3.0

# Protein Graph Construction
graphein>=1.7.0
biopython>=1.81
networkx>=3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress and Logging
tqdm>=4.65.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0
